requirements.txt
Django>=1.3,<1.4
MySQL-python
python-memcached
django-extensions
redis
django-celery
gunicorn
